test_app/requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
protobuf==5.28.3
PyMuPDF==1.24.13
PyYAML==6.0.2
sniffio==1.3.1